+++ src/main/java/com/qa/CVConsumer/CvConsumerApplication.java
@@ -1,0 +1,12 @@
+package com.qa.CVConsumer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CvConsumerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CvConsumerApplication.class, args);
+	}
+}
