--- src/main/java/com/qa/consumer/persistence/domain/Trainee.java
@@ -1,33 +1,0 @@
-package com.qa.consumer.persistence.domain;
-
-public class Trainee {
-
-	private String firstName;
-	private String lastName;
-	private String email;
-
-	public String getFirstName() {
-		return firstName;
-	}
-
-	public void setFirstName(String firstName) {
-		this.firstName = firstName;
-	}
-
-	public String getLastName() {
-		return lastName;
-	}
-
-	public void setLastName(String lastName) {
-		this.lastName = lastName;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-}
