--- src/main/java/com/qa/consumer/util/Constants.java
@@ -1,24 +1,0 @@
-package com.qa.consumer.util;
-
-public class Constants {
-
-	private Constants() {
-	}
-
-	public static final String CV_QUEUED_MESSAGE = "File placed on queue succesfully";
-
-	public static final String MALFORMED_REQUEST_MESSAGE = "Queue request malformed";
-
-	public static final String CV_ADDED_MESSAGE = "CV added succesfully";
-
-	public static final String CV_NOT_FOUND_MESSAGE = "CV not found";
-
-	public static final String CV_DELETED_MESSAGE = "CV deleted succesfully";
-
-	public static final String CV_UPDATED_MESSAGE = "CV updated succesfully";
-
-	public static final String INCOMING_QUEUE_NAME = "IncomingQueue";
-
-	public static final String OUTGOING_QUEUE_NAME = "OutgoingQueue";
-
-}
