--- src/main/resources/application.properties
@@ -1,13 +1,0 @@
-incomingQueueName = incomingQueue
-outgoingQueueName = outgoingQueue
-
-server.port = 8080
-
-spring.activemq.user=admin
-spring.activemq.password=admin
-spring.activemq.broker-url=tcp://localhost:61616
-
-spring.data.mongodb.database=CVStore
-spring.data.mongodb.host=localhost
-spring.data.mongodb.port=27017
-spring.main.allow-bean-definition-overriding=true
