+++ src/test/java/com/qa/CVConsumer/CvConsumerApplicationTests.java
@@ -1,0 +1,16 @@
+package com.qa.CVConsumer;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class CvConsumerApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
